requirements.txt
absl-py==1.2.0
accelerate==0.24.1
astroid==2.15.6
asttokens==2.4.0
astunparse==1.6.3
backcall==0.2.0
CacheControl==0.13.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.16.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
comm==0.1.4
cryptography==41.0.4
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
dnspython==2.4.2
exceptiongroup==1.1.3
executing==1.2.0
filelock==3.12.2
firebase-admin==6.2.0
flake8==6.1.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.0
flatbuffers==1.12
fonttools==4.34.4
fsspec==2023.9.2
gast==0.4.0
google-api-core==2.12.0
google-api-python-client==2.101.0
google-auth==2.23.2
google-auth-httplib2==0.1.1
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
greenlet==1.1.3
grpcio==1.59.0
grpcio-status==1.59.0
h5py==3.7.0
httplib2==0.22.0
huggingface-hub==0.17.3
idna==3.3
importlib-metadata==4.12.0
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.15.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.1.0
jupyter_client==8.3.1
jupyter_core==5.3.1
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
libclang==14.0.6
loguru==0.7.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mpmath==1.3.0
msgpack==1.0.7
mysql==0.0.3
mysql-connector-python==8.1.0
mysqlclient==2.1.1
nest-asyncio==1.5.8
networkx==3.1
nltk==3.8.1
numpy==1.23.1
oauthlib==3.2.0
opt-einsum==3.3.0
ortools==9.7.2996
packaging==21.3
pandas==1.4.3
parso==0.8.3
patsy==0.5.3
pgmpy==0.1.23
pickleshare==0.7.5
pika==1.3.2
Pillow==9.2.0
pinecone-client==2.2.4
platformdirs==3.10.0
pluggy==1.3.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.3
psutil==5.9.5
PuLP==2.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
Pygments==2.16.1
PyJWT==2.8.0
pylint==2.17.5
pyparsing==3.0.9
pytest==7.4.2
pytest-dependency==0.5.1
python-dateutil==2.8.2
pytz==2022.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.8.8
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.0
scikit-learn==1.1.1
scipy==1.9.0
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.41
stack-data==0.6.2
statsmodels==0.14.0
sympy==1.12
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tokenizers==0.14.1
tomli==2.0.1
tomlkit==0.12.1
torch==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
transformers==4.35.0
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
wcwidth==0.2.6
Werkzeug==2.2.2
win32-setctime==1.1.0
wrapt==1.14.1
zipp==3.8.1
